Don't display pending validations in the host status popover
Display only failing validations in host status popover. Display pending
validations only if there are no failing ones (should not really ever happen).
The reason for this is that pending validations are always pending due to
another validation being failed.

The validation group status is 'Pending input' if there are any pending validations.
This is to better communicate that validations are failing due to us waiting for
user input.