BZ-1928653 Fixes flooding sentry when there is and API error
This PR modifies the polling mechanism so it doesn't continues fetching
when there is an API error.
Also, the 'try again' action in the `ErrorState` component will trigger a full page reload only when the received HTTP status code is 401 Unauthorized, this is done in order to force OCM to send the users to the login page so they can renew their token.

Signed-off-by: Jonathan Kilzi <jkilzi@redhat.com>