BZ-1953528 Don't call to PATCH API if the for m hasn't changed
Before this PR we were calling the PATCH API on every transition from the "Cluster Details" to the "Host Discovery" (first to second step) even though the form remains pristine.
For the rest of the steps there is no PATCH API call when the form is pristine. This PR prevents calling the PATCH API when the form is pristine during the first transition only.

Signed-off-by: Jonathan Kilzi <jkilzi@redhat.com>