Use custom resolve conditions (#2063)
Using 'development' as the package exports condition clashes with the settings in the uhc-portal where Webpack is used and the default condition is 'development'. Since that project's Webpack config is poorly configured, we are not able to tell them to import from the default condition without triggering a major refactor on their side. This change settles the issue with minimal impact on their side.