MGMT-15140: Do not report delayed CIM config as an error
CIM initial configuration might take long time depending
on the customer environment.
The UI can not detect an error, we just report "unusual"
condition if the process takes longer than a threshold,
leaving it up to the user to investigate and decide based
on the provided guiding.

Prior this patch, we reported the "delayed progress" as
an error (using the `danger` alert).
By this change, it's less aggressive and less
confusing `info` message instead.

The timeout for "potential hanging" message is
increased from 3 to 5 minutes.