Update setAuthInterceptor to be indempotent
Today if we call setAuthInterceptor multiple times, it's not working because we reuse previous client and clientWithoutCaseConverter.
Config url in withAssistedInstallerBasePath for example is applied multiple times.

Normally we call setAuthInterceptor only one time. But with recent module federated code, we call setAuthInterceptor every time we load the component.
We also remove the job template created earlier