MGMT-19409: Disable OpenShift AI when LVM is enabled
The OpenShift AI operator requires ODF, which in turn is incompatible
with the LVM operator. But the UI isn't aware of that. As a result when
the user selects the LVM operator and then tries to select the OpenShift
AI operator the UI allows it, but the backend returns an error. To avoid
that this patch changes the UI so that when LVM is selected the
OpenShift AI option be automatically disabled.

Related: https://issues.redhat.com/browse/MGMT-19409
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>